Phase 11: busybox sh を動かす (echo / 複数コマンド)
主な修正:
- Cpu64: XCHG rax, r8 (49 90) を NOP として誤処理していたのを修正
- Cpu64: 0xC7 で 0x66 operand-size プレフィックス無視を修正 (imm32→imm16)
- Cpu64: Grp4 (FE) INC/DEC r/m8 を実装
- Syscall: sys_uname の int 切り詰め残党を long へ修正
- SyscallAmd64: getcwd (#79) を新設

debug.mk に busybox 系ターゲット (bb-sh / bb-ls / bb-cat / bb-echo /
bb-grep / bb-prep) を追加。許可ダイアログを回避してパイプ + 日本語パスを
使わずにテスト可能に。

回帰テスト: 43 PASS / 0 FAIL / 1 SKIP (sys_chmod64 の WSL DrvFs 制約のみ)

\`busybox sh -c "echo hello from sh"\` 等が動作。for/if 内の変数展開や
fork 経由の出力には未解決の問題あり (Phase 12 候補)。

+++ src/main/java/emulin/Syscall.java
@@ -1116,7 +1116,7 @@ public class Syscall extends EmuSocket
   }
   long sys_uname( long bx, long cx, long dx, long si, long di )       {
     final int SYS_NMLN = 65;
-    int address = (int)bx;
+    long address = bx;
     int ret = 0;
     InetAddress iaddr = null;
     String s;
